Expand base [de]serialize tests to complex types
